+++ index.html
@@ -6,7 +6,7 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кексобукинг</title>
-  <script src="js/main.js" defer></script>
+  <script type="module" src="js/main.js" defer></script>
 </head>
 <body>
