+++ js/data.js
@@ -1,0 +1,33 @@
+export const ADVENTS_COUNTS = 10,
+  TYPES = [
+    'palace',
+    'flat',
+    'house',
+    'bungalow',
+    'hotel',
+  ],
+  FEATURES_LIST = [
+    'wifi',
+    'dishwasher',
+    'parking',
+    'washer',
+    'elevator',
+    'conditioner',
+  ],
+  DESCRIPTIONS = [
+    'Отличное жилище',
+    'Неплохое помещение',
+    'Хорошее место для ночлега',
+  ],
+  BASEDIR = 'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/',
+  PHOTOS = [
+    'duonguyen-8LrGtIxxa4w.jpg',
+    'brandon-hoogenboom-SNxQGWxZQi0.jpg',
+    'claire-rendall-b6kAwr1i0Iw.jpg',
+  ];
+export const adventTitle = 'Объявление',
+  [minPrice, maxPrice] = [1, 50000],
+  [minRooms, maxRooms] = [1, 10],
+  [minGuests, maxGuests] = [1, 10],
+  [startTime, endTime] = [12, 14],
+  [latMin, latMax, lngMin, lngMax, accu] = [35.65000, 35.70000, 139.70000, 139.80000, 6];
