--- js/utils/get-random-positive-float.js
@@ -1,9 +1,0 @@
-// Функция взята из интернета и доработана
-// Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
-
-// function getRandomPositiveFloat (a, b, digits = 1) {
-//   const lower = Math.min(Math.abs(a), Math.abs(b));
-//   const upper = Math.max(Math.abs(a), Math.abs(b));
-//   const result = Math.random() * (upper - lower) + lower;
-//   return result.toFixed(digits);
-// }
